chore: Update npm dependencies and add Flowbite library
This commit updates the npm dependencies in the package.json file to include the latest stable versions of "@types/prettier", "prettier", "react", and "react-dom". Additionally, it adds the "flowbite" and "flowbite-react" libraries as dependencies. The purpose of this update is to ensure compatibility with the latest versions and to leverage the features provided by the Flowbite library.