chore: Add react-slick library and update npm dependencies
This commit adds the "react-slick" library as a dependency in the package.json file. It also updates the npm dependencies to include the latest stable versions of "@types/react-slick", "@typescript-eslint/eslint-plugin", and "@typescript-eslint/parser". The purpose of these changes is to incorporate the react-slick library for carousel functionality and ensure compatibility with the latest versions of the dependencies.